+++ lib/src/environment/environment_service_web.dart
@@ -58,6 +58,12 @@ class EnvironmentService implements EnvironmentServiceInterface {
     if (value != null) {
       switch (key) {
         case otelResourceAttributes:
+        case otelDartLogMetrics:
+          return const String.fromEnvironment(otelDartLogMetrics);
+        case otelDartLogSpans:
+          return const String.fromEnvironment(otelDartLogSpans);
+        case otelDartLogExport:
+          return const String.fromEnvironment(otelDartLogExport);
         case otelPropagators:
         case otelExporterOtlpHeaders:
         case otelExporterOtlpTracesHeaders:
@@ -100,12 +106,6 @@ class EnvironmentService implements EnvironmentServiceInterface {
         return const String.fromEnvironment(otelTracesSamplerArg);
 
       // Logging Configuration
-      case otelLogMetrics:
-        return const String.fromEnvironment(otelLogMetrics);
-      case otelLogSpans:
-        return const String.fromEnvironment(otelLogSpans);
-      case otelLogExport:
-        return const String.fromEnvironment(otelLogExport);
 
       // General OTLP Configuration
       case otelExporterOtlpEndpoint: