+++ lib/src/trace/tracer_provider.dart
@@ -227,7 +227,7 @@ class TracerProvider implements APITracerProvider {
         if (resource != null) {
           OTelLog.debug('Resource attributes:');
           resource!.attributes.toList().forEach((attr) {
-            if (attr.key == 'tenant_id' || attr.key == 'service.name') {
+            if (attr.key == 'service.name') {
               OTelLog.debug('  ${attr.key}: ${attr.value}');
             }
           });